+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Change Log
-- Update `ZSQuotesDiagnosticProvider` to underline curly quotes mark with red underline
+- Update `detectStraightQuotes` to underline curly quotes mark with red underline
 - Up to version 0.1.12