+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Change Log
-- Update `detectStraightQuotes` to underline curly quotes mark with red underline
-- Up to version 0.1.12
+- Remove `line.includes('*/-') || line.includes('/-*')` in `detectStraightQuotes`
+- Up to version 0.1.13